use lists for data pointers for each identity
