+++ src/App.tsx
@@ -17,14 +17,7 @@
 import '../shim';
 import 'utils/iconLoader';
 import * as React from 'react';
-import {
-	AppState,
-	AppStateStatus,
-	StatusBar,
-	StyleSheet,
-	View,
-	YellowBox
-} from 'react-native';
+import { StatusBar, StyleSheet, View, YellowBox } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { Provider as UnstatedProvider } from 'unstated';
 import { MenuProvider } from 'react-native-popup-menu';
@@ -36,7 +29,7 @@ import {
 	ScreenStack
 } from './screens';
 
-import { useSeedRef } from 'utils/seedRefHooks';
+import { SeedRefStore } from 'stores/SeedRefStore';
 import colors from 'styles/colors';
 import '../ReactotronConfig';
 import { AppProps, getLaunchArgs } from 'e2e/injections';
@@ -59,26 +52,6 @@ export default function App(props: AppProps): React.ReactElement {
 
 	const [policyConfirmed, setPolicyConfirmed] = React.useState<boolean>(false);
 	const [dataLoaded, setDataLoaded] = React.useState<boolean>(false);
-	const [appState, setAppState] = React.useState<AppStateStatus>(
-		AppState.currentState
-	);
-	const { destroySeedRef, isSeedRefValid } = useSeedRef();
-
-	React.useEffect(() => {
-		const _handleAppStateChange = (nextAppState: AppStateStatus): void => {
-			if (nextAppState.match(/inactive|background/) && appState === 'active' && isSeedRefValid) {
-				destroySeedRef().then(() =>
-					console.log('is seed ref valid: ', isSeedRefValid)
-				);
-			}
-			setAppState(nextAppState);
-		};
-		AppState.addEventListener('change', _handleAppStateChange);
-
-		return (): void => {
-			AppState.removeEventListener('change', _handleAppStateChange);
-		};
-	}, [appState, destroySeedRef, isSeedRefValid]);
 
 	React.useEffect(() => {
 		const loadPolicyConfirmationAndMigrateData = async (): Promise<void> => {
@@ -122,14 +95,16 @@ export default function App(props: AppProps): React.ReactElement {
 
 	return (
 		<SafeAreaProvider>
-			<UnstatedProvider>
-				<MenuProvider backHandler={true}>
-					<StatusBar barStyle="light-content" backgroundColor={colors.bg} />
-					<GlobalStateContext.Provider value={globalContext}>
-						<NavigationContainer>{renderStacks()}</NavigationContainer>
-					</GlobalStateContext.Provider>
-				</MenuProvider>
-			</UnstatedProvider>
+			<SeedRefStore>
+				<UnstatedProvider>
+					<MenuProvider backHandler={true}>
+						<StatusBar barStyle="light-content" backgroundColor={colors.bg} />
+						<GlobalStateContext.Provider value={globalContext}>
+							<NavigationContainer>{renderStacks()}</NavigationContainer>
+						</GlobalStateContext.Provider>
+					</MenuProvider>
+				</UnstatedProvider>
+			</SeedRefStore>
 		</SafeAreaProvider>
 	);
 }
