+++ src/components/IdentitiesSwitch.tsx
@@ -23,7 +23,6 @@ import ButtonIcon from './ButtonIcon';
 import Separator from './Separator';
 import TransparentBackground from './TransparentBackground';
 
-import { useSeedRef } from 'utils/seedRefHooks';
 import { RootStackParamList } from 'types/routes';
 import AccountsStore from 'stores/AccountsStore';
 import testIDs from 'e2e/testIDs';
@@ -35,8 +34,7 @@ import {
 	getSeedPhrase,
 	navigateToLegacyAccountList,
 	resetNavigationTo,
-	resetNavigationWithNetworkChooser,
-	unlockSeedPhrase
+	resetNavigationWithNetworkChooser
 } from 'utils/navigationHelpers';
 import { Identity } from 'types/identityTypes';
 
@@ -48,7 +46,6 @@ function IdentitiesSwitch({
 	const navigation: StackNavigationProp<RootStackParamList> = useNavigation();
 	const [visible, setVisible] = useState(false);
 	const { currentIdentity, identities } = accounts.state;
-	const { isSeedRefValid } = useSeedRef();
 	// useEffect(() => {
 	// 	const firstLogin: boolean = identities.length === 0;
 	// 	if (currentIdentity === null && !firstLogin) {
@@ -74,7 +71,7 @@ function IdentitiesSwitch({
 	): Promise<void> => {
 		await accounts.selectIdentity(identity);
 		setVisible(false);
-		if (screenName === 'AccountNetworkChooser') {
+		if (screenName === 'Main') {
 			resetNavigationTo(navigation, screenName, params);
 		} else if (screenName === 'IdentityBackup') {
 			const seedPhrase = await getSeedPhrase(navigation);
@@ -135,10 +132,7 @@ function IdentitiesSwitch({
 				<ButtonIcon
 					title={currentIdentityTitle}
 					onPress={(): Promise<void> =>
-						onIdentitySelectedAndNavigate(
-							currentIdentity,
-							'AccountNetworkChooser'
-						)
+						onIdentitySelectedAndNavigate(currentIdentity, 'Main')
 					}
 					iconType="antdesign"
 					iconName="user"
@@ -201,7 +195,7 @@ function IdentitiesSwitch({
 				}
 				title={title}
 				onPress={(): Promise<void> =>
-					onIdentitySelectedAndNavigate(identity, 'AccountNetworkChooser')
+					onIdentitySelectedAndNavigate(identity, 'Main')
 				}
 				key={identity.encryptedSeed}
 				iconType="antdesign"
