+++ src/modules/main/styles.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+	onboardingWrapper: {
+		alignItems: 'center',
+		flexDirection: 'row',
+		flexWrap: 'wrap'
+	},
+	scrollContent: {
+		flex: 1,
+		justifyContent: 'center',
+		padding: 16,
+		paddingBottom: 100
+	}
+});
+
+export default styles;
