+++ src/screens/IdentityNew.tsx
@@ -37,6 +37,7 @@ import ScreenHeading from 'components/ScreenHeading';
 import KeyboardScrollView from 'components/KeyboardScrollView';
 import { brainWalletAddress } from 'utils/native';
 import { debounce } from 'utils/debounce';
+import { useGenerateSeedRef } from 'utils/seedRefHooks';
 
 function IdentityNew({
 	accounts,
@@ -48,6 +49,7 @@ function IdentityNew({
 	const [isRecover, setIsRecover] = useState(isRecoverDefaultValue);
 	const [isSeedValid, setIsSeedValid] = useState(defaultSeedValidObject);
 	const [seedPhrase, setSeedPhrase] = useState('');
+	const generateNewSeed = useGenerateSeedRef();
 
 	useEffect((): (() => void) => {
 		const clearNewIdentity = (): void =>
@@ -76,9 +78,13 @@ function IdentityNew({
 		const pin = await setPin(navigation);
 		try {
 			if (isSeedValid.bip39) {
-				await accounts.saveNewIdentity(seedPhrase.trimEnd(), pin);
+				await accounts.saveNewIdentity(
+					seedPhrase.trimEnd(),
+					pin,
+					generateNewSeed
+				);
 			} else {
-				await accounts.saveNewIdentity(seedPhrase, pin);
+				await accounts.saveNewIdentity(seedPhrase, pin, generateNewSeed);
 			}
 			setSeedPhrase('');
 			navigateToNewIdentityNetwork(navigation);
