+++ src/types/routes.ts
@@ -6,7 +6,7 @@ export type RootStackParamList = {
 	AccountEdit: undefined;
 	AccountPin: { isNew: boolean } | undefined;
 	AccountNew: undefined;
-	AccountNetworkChooser: { isNew: boolean } | undefined;
+	Main: { isNew: boolean } | undefined;
 	AccountUnlockAndSign: { next: string };
 	AccountUnlock: { next: string; onDelete: () => any };
 	IdentityBackup: { isNew: true } | { isNew: false; seedPhrase: string };
