+++ test/e2e/specs/signing.spec.ts
@@ -32,7 +32,7 @@ import { ScanTestRequest } from 'e2e/mockScanRequests';
 import testIDs from 'e2e/testIDs';
 
 const {
-	AccountNetworkChooser,
+	Main,
 	PathDetail,
 	SecurityHeader,
 	TxDetails,
@@ -93,16 +93,16 @@ describe('Signing test', () => {
 		it('generate Kovan account', async () => {
 			await tapBack();
 			const kovanNetworkButtonIndex =
-				AccountNetworkChooser.networkButton + EthereumNetworkKeys.KOVAN;
-			await testTap(testIDs.AccountNetworkChooser.addNewNetworkButton);
+				Main.networkButton + EthereumNetworkKeys.KOVAN;
+			await testTap(testIDs.Main.addNewNetworkButton);
 			await testScrollAndTap(
 				kovanNetworkButtonIndex,
-				testIDs.AccountNetworkChooser.chooserScreen
+				testIDs.Main.chooserScreen
 			);
 			await testUnlockPin(pinCode);
 			await testVisible(PathDetail.screen);
 			await tapBack();
-			await testExist(AccountNetworkChooser.chooserScreen);
+			await testExist(Main.chooserScreen);
 		});
 
 		it('should sign transactions', async () => {
