+++ test/e2e/utils.ts
@@ -17,20 +17,13 @@
 import { expect, element, by, device } from 'detox';
 
 import testIDs from './testIDs';
-const {
-	IdentityPin,
-	IdentityNew,
-	AccountNetworkChooser,
-	PathDetail,
-	PathsList
-} = testIDs;
+const { IdentityPin, IdentityNew, Main, PathDetail, PathsList } = testIDs;
 
 export const mockIdentityName = 'mockIdentity';
 export const mockSeedPhrase =
 	'ability cave solid soccer gloom thought response hard around minor want welcome';
 export const pinCode = '000000';
-const substrateNetworkButtonIndex =
-	AccountNetworkChooser.networkButton + 'kusama';
+const substrateNetworkButtonIndex = Main.networkButton + 'kusama';
 
 export const testTap = async (buttonId: string): Promise<Detox.Actions<any>> =>
 	await element(by.id(buttonId)).tap();
@@ -93,10 +86,10 @@ export const testUnlockPin = async (inputPin: string): Promise<void> => {
 export const testSetUpDefaultPath = async (): Promise<void> => {
 	await testInput(IdentityPin.setPin, pinCode);
 	await testInputWithDone(IdentityPin.confirmPin, pinCode);
-	await testVisible(AccountNetworkChooser.chooserScreen);
+	await testVisible(Main.chooserScreen);
 	await testScrollAndTap(
 		substrateNetworkButtonIndex,
-		testIDs.AccountNetworkChooser.chooserScreen
+		testIDs.Main.chooserScreen
 	);
 	await testUnlockPin(pinCode);
 	await testVisible(PathDetail.screen);
@@ -116,7 +109,7 @@ export const launchWithScanRequest = async (
 
 export const testRecoverIdentity = (): void => {
 	it('recover a identity with seed phrase', async () => {
-		await testTap(AccountNetworkChooser.recoverButton);
+		await testTap(Main.recoverButton);
 		await testVisible(IdentityNew.seedInput);
 		await testInput(IdentityNew.nameInput, mockIdentityName);
 		await element(by.id(IdentityNew.seedInput)).typeText(mockSeedPhrase);
